+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nix run issue [#272]
 - Range not displayed when fetch fails [#276]
 - Blank lines and spaces in `text/plain` parts [#280]
+- Watch command [#271]
 
 ### Removed
 
@@ -373,4 +374,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [#272]: https://github.com/soywod/himalaya/issues/272
 [#273]: https://github.com/soywod/himalaya/issues/273
 [#276]: https://github.com/soywod/himalaya/issues/276
+[#271]: https://github.com/soywod/himalaya/issues/271
 [#280]: https://github.com/soywod/himalaya/issues/280